split print hook to facilitate tests
