+++ alive_progress/core/progress.py
@@ -9,6 +9,7 @@ from itertools import chain, islice, repeat
 from .calibration import calibrated_fps
 from .configuration import config_handler
 from .logging_hook import install_logging_hook, uninstall_logging_hook
+from .hook_manager import create_print_hook
 from .timing import gen_simple_exponential_smoothing_eta, to_elapsed_text, to_eta_text
 from .utils import clear_traces, hide_cursor, render_title, sanitize_text_marking_wide_chars, \
     show_cursor, terminal_columns
@@ -116,10 +117,6 @@ def alive_bar(total=None, title=None, calibrate=None, **options):
 
         run.last_line_len = line_len
 
-    def flush_buffer():
-        if print_buffer:
-            print()
-
     def set_text(message):
         run.text = sanitize_text_marking_wide_chars(message)
 
@@ -156,26 +153,6 @@ def alive_bar(total=None, title=None, calibrate=None, **options):
                 set_text(text)
             return run.count
     bar.text = set_text
-
-    def print_hook(part):
-        if part != '\n':
-            # this will generate a sequence of lines interspersed with None, which will later
-            # be rendered as the indent filler to align additional lines under the same header.
-            gen = chain.from_iterable(zip(repeat(None), part.splitlines(True)))
-            print_buffer.extend(islice(gen, 1, None))
-        else:
-            header = header_template.format(run.count)
-            nested = ''.join(line or ' ' * len(header) for line in print_buffer)
-            with print_lock:
-                clear_traces()
-                sys.__stdout__.write('{}{}\n'.format(header, nested))
-            print_buffer[:] = []
-
-    print_buffer, print_lock = [], threading.Lock()
-    header_template = 'on {}: ' if config.enrich_print else ''
-    print_hook.write = print_hook
-    print_hook.flush = lambda: None
-    print_hook.isatty = sys.__stdout__.isatty
 
     def start_monitoring(offset=0.):
         hide_cursor()
